The return of AUTO programming? 23pts
