+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/INTO_THE_DEEP_BOT/_/Autonomous/Blue/BLUE_NETZONE_LowBasket.java
@@ -1,12 +1,12 @@
 package org.firstinspires.ftc.teamcode.INTO_THE_DEEP_BOT._.Autonomous.Blue;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
-import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 import org.firstinspires.ftc.teamcode.INTO_THE_DEEP_BOT._.Autonomous.AutonomousPLUS;
-@Autonomous(group = "Basic", name = "Blue Basic Net Zone")
-public class BlueBasicNetZone extends AutonomousPLUS {
+
+@Autonomous(group = "Basic", name = "Net Zone Basket")
+public class BLUE_NETZONE_LowBasket extends AutonomousPLUS {
 
     public String currentPosition;
     public String target;
@@ -14,33 +14,28 @@ public class BlueBasicNetZone extends AutonomousPLUS {
     public void runOpMode() {
 
         super.runOpMode();
+
+
         waitForStart();
         telemetry.addData(currentPosition,"Start position");
-
         robot.lifty.setPower(0);
         robot.tempOutakePos("UP");
         robot.intakePosition("UP");
-        speed = .3;
+        speed = .4;
         robot.outakeclawOpenClose("CLOSED");
-        moveRobotForward(500, 0);
-        moveRobotRight(1000, 0);
-        robot.lifty.setPower(1);
-        robot.lifty.setTargetPosition(1500);
-        robot.lifty.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        sleep(1000);
-        moveRobotForward(500, 2);
+        //Movement Starts
+        moveRobotForward(900,2);
+        turnRobotLeft(1250,2);
+        moveRobotForward(950,2);
         robot.lifty.setPower(1);
         robot.lifty.setTargetPosition(2500);
         robot.lifty.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        prepareNextAction(1000);
+        sleep(3000);
+        moveRobotForward(150,2);
         robot.outakeclawOpenClose("OPEN");
-        prepareNextAction(2);
-        moveRobotBackward(200,2);
-        /*robot.lifty.setTargetPosition(1228);
-        robot.lifty.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        moveRobotLeft(1000, 2);
-        turnRobotRight(900, 2);
-        moveRobotLeft(900, 2);
-        moveRobotForward(200, 2);*/
+        sleep(1000);
+        speed = 1;
+        turnRobotLeft(15,2);
+
     }
 }
