+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/INTO_THE_DEEP_BOT/_/Autonomous/Blue/OLD_LowBasket.java
@@ -1,12 +1,13 @@
 package org.firstinspires.ftc.teamcode.INTO_THE_DEEP_BOT._.Autonomous.Blue;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
+import com.qualcomm.robotcore.eventloop.opmode.Disabled;
 import com.qualcomm.robotcore.hardware.DcMotor;
 
 import org.firstinspires.ftc.teamcode.INTO_THE_DEEP_BOT._.Autonomous.AutonomousPLUS;
-
-@Autonomous(group = "Basic", name = "Blue Basic Net Zone (Basket) Better")
-public class BlueBasicNetZoneBasketBetter extends AutonomousPLUS {
+@Disabled
+@Autonomous(group = "Basic", name = "Blue Basic Net Zone (Basket)")
+public class OLD_LowBasket extends AutonomousPLUS {
 
     public String currentPosition;
     public String target;
@@ -18,24 +19,28 @@ public class BlueBasicNetZoneBasketBetter extends AutonomousPLUS {
 
         waitForStart();
         telemetry.addData(currentPosition,"Start position");
+
         robot.lifty.setPower(0);
         robot.tempOutakePos("UP");
-        robot.intakePosition("UP");
-        speed = .4;
+        robot.slidesIn();
+        speed = .65;
         robot.outakeclawOpenClose("CLOSED");
         //Movement Starts
-        moveRobotForward(900,2);
-        turnRobotLeft(1250,2);
-        moveRobotForward(950,2);
+        moveRobotForward(200,2);
+        turnRobotLeft(850,2);
+        robot.lifty.setPower(1);
+        robot.lifty.setTargetPosition(1500);
+        robot.lifty.setMode(DcMotor.RunMode.RUN_TO_POSITION);
+        sleep(1500);
+        moveRobotForward(550,2);
+        moveRobotRight(450,2);
+        turnRobotLeft(400,2);
         robot.lifty.setPower(1);
         robot.lifty.setTargetPosition(2500);
         robot.lifty.setMode(DcMotor.RunMode.RUN_TO_POSITION);
-        sleep(3000);
-        moveRobotForward(150,2);
+        sleep(1500);
+        moveRobotForward(360,2);
         robot.outakeclawOpenClose("OPEN");
-        sleep(1000);
-        speed = 1;
-        turnRobotLeft(15,2);
-
+        sleep(4000);
     }
 }
